chore(jaxrs,security): Move JAX-RS security config to build-time
